Move SpecialField class to particular package in model
