+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldDropDown.java
@@ -12,10 +12,10 @@ import javax.swing.JPopupMenu;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.specialfields.SpecialField;
+import net.sf.jabref.model.entry.specialfields.SpecialField;
 import net.sf.jabref.logic.specialfields.SpecialFieldLocalization;
 import net.sf.jabref.logic.util.OS;
-import net.sf.jabref.model.entry.SpecialFieldValue;
+import net.sf.jabref.model.entry.specialfields.SpecialFieldValue;
 
 import com.jgoodies.looks.HeaderStyle;
 import com.jgoodies.looks.Options;
