+++ src/main/java/net/sf/jabref/gui/util/comparator/RankingFieldComparator.java
@@ -4,7 +4,7 @@ import java.util.Comparator;
 import java.util.Optional;
 
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.model.entry.SpecialFields;
+import net.sf.jabref.model.entry.specialfields.SpecialFields;
 
 /**
  * Comparator that handles the ranking icon column
