+++ src/test/java/net/sf/jabref/logic/specialfields/SpecialFieldsUtilsTest.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import net.sf.jabref.model.FieldChange;
 import net.sf.jabref.model.entry.BibEntry;
 
+import net.sf.jabref.model.entry.specialfields.SpecialField;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
